requirements.txt
blinker==1.9.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==45.0.7
dotenv==0.9.9
Flask==3.1.2
flask-cors==6.0.1
Flask-SQLAlchemy==3.1.1
fpdf==1.7.2
greenlet==3.2.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.1
MarkupSafe==3.0.2
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2==2.9.10
pycparser==2.23
PyPDF2==3.0.1
pypdfium2==4.30.0
python-docx==1.2.0
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
flasgger==0.9.7.1
Werkzeug==3.1.3
gunicorn==20.1.0
flask_swagger_ui==5.21.0